# swagger.yaml
openapi: 3.0.0
info:
  title: Ethio Social Media Platform API
  description: " This is the official API documentation for Ethio Social Media Platform — a blockchain-based social media platform designed to support secure, transparent, and tokenized gifiting system.\r\n\r\nKey features of the platform include:\r\n- User registration and profile management\r\n- Token-based (Stars) gifting and premium content access\r\n- Wallet integration and token purchase using the Token Sale smart contract\r\n- Reward distribution to content creators\r\n- Content moderation for hate or abusive speech, including Amharic language support\r\n- Integration with smart contracts deployed on the Polygon/Sepolia testnet\r\n\r\nAll API endpoints are designed to interact with the backend services that support user management, transaction logging, content control, and blockchain event handling."
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        authorId:
          type: string
          format: uuid
        likes:
          type: integer
        comments:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PostReport:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contentId:
          type: string
          format: uuid
        reason:
          type: string
        reportedBy:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, RESOLVED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /social/posts:
    get:
      tags:
        - Social
      summary: Get all posts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: authorId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  total:
                    type: integer

    post:
      tags:
        - Social
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
  /social/posts/{id}:
    get:
      tags:
        - Social
      summary: Get post by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found

    patch:
      tags:
        - Social
      summary: Update post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

    delete:
      tags:
        - Social
      summary: Delete post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Post deleted successfully
  /social/posts/{id}/report:
    post:
      tags:
        - Social
      summary: Report a post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
      responses:
        '201':
          description: Post reported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /social/posts/{id}/reports:
    get:
      tags:
        - Social
      summary: Get post reports
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of post reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostReport'
...